merge: also validate conflicting external ids and cross-links between merged entities for dataadmin merges
